Separate out interface for DocumentOperationExecutor
